tpkg: renamed to toitpkg
related https://github.com/toitlang/tpkg/pull/2
