+++ completers/toitpkg_completer/cmd/completion_bash.go
