+++ completers/toitpkg_completer/cmd/completion_fish.go
