+++ completers/toitpkg_completer/cmd/completion_zsh.go
