+++ completers/toitpkg_completer/cmd/pkg_clean.go
