+++ completers/toitpkg_completer/cmd/pkg_registry.go
