+++ completers/toitpkg_completer/cmd/pkg_registry_list.go
