+++ completers/toitpkg_completer/cmd/pkg_search.go
