+++ completers/toitpkg_completer/cmd/pkg_uninstall.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/rsteube/carapace"
-	"github.com/rsteube/carapace-bin/completers/tpkg_completer/cmd/action"
+	"github.com/rsteube/carapace-bin/completers/toitpkg_completer/cmd/action"
 	"github.com/spf13/cobra"
 )
