+++ completers/toitpkg_completer/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/rsteube/carapace-bin/completers/toitpkg_completer/cmd"
+
+func main() {
+	cmd.Execute()
+}
