--- completers/tpkg_completer/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/rsteube/carapace-bin/completers/tpkg_completer/cmd"
-
-func main() {
-	cmd.Execute()
-}
